python/requirements.txt
psycopg2-binary==2.9.7
pandas==2.1.1
faker==19.6.2
numpy==1.24.3
python-dotenv==1.0.0